Remove cloud artifacts and update config
This repo is a fork of hubverse-org/example-complex-forecast-hub that
was subsequently removed from the fork network (i.e., it's now a standalone
repo).

We wanted a repo to practice onboarding a hub to the cloud, so this
commit removes the cloud section of admin.json and the GitHub worfklow
that synchronizes data to S3.